requirements.txt
streamlit==1.12.0
pandas
numpy
tensorflow
Pillow